requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
amqp==5.2.0
    # via kombu
asgiref==3.8.1
    # via django
async-timeout==5.0.1
    # via redis
autopep8==2.3.1
    # via django-silk
billiard==4.2.1
    # via celery
blessed==1.20.0
    # via curtsies
bpython==0.24
    # via -r requirements.in
build==1.2.2.post1
    # via pip-tools
celery[redis]==5.4.0
    # via
    #   -r requirements.in
    #   django-celery-beat
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cron-descriptor==1.4.5
    # via django-celery-beat
curtsies==0.4.2
    # via bpython
cwcwidth==0.1.9
    # via
    #   bpython
    #   curtsies
django==4.2.16
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-redis
    #   django-silk
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-celery-beat==2.7.0
    # via -r requirements.in
django-redis==5.4.0
    # via -r requirements.in
django-silk==5.3.0
    # via -r requirements.in
django-timezone-field==7.0
    # via django-celery-beat
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
drf-yasg==1.21.8
    # via -r requirements.in
factory-boy==3.3.1
    # via -r requirements.in
faker==30.8.2
    # via factory-boy
gprof2dot==2024.6.6
    # via django-silk
greenlet==3.1.1
    # via bpython
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-yasg
kombu==5.4.2
    # via celery
packaging==24.1
    # via
    #   build
    #   drf-yasg
pillow==11.0.0
    # via -r requirements.in
pip-tools==7.4.1
    # via -r requirements.in
prompt-toolkit==3.0.48
    # via click-repl
psycopg2==2.9.10
    # via -r requirements.in
pycodestyle==2.12.1
    # via autopep8
pygments==2.18.0
    # via bpython
pyjwt==2.9.0
    # via djangorestframework-simplejwt
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   faker
    #   python-crontab
pytz==2024.2
    # via drf-yasg
pyxdg==0.28
    # via bpython
pyyaml==6.0.2
    # via drf-yasg
redis==5.2.0
    # via
    #   celery
    #   django-redis
requests==2.32.3
    # via bpython
six==1.16.0
    # via
    #   blessed
    #   python-dateutil
sqlparse==0.5.1
    # via
    #   django
    #   django-silk
tomli==2.0.2
    # via
    #   autopep8
    #   build
    #   pip-tools
typing-extensions==4.12.2
    # via
    #   asgiref
    #   faker
tzdata==2024.2
    # via
    #   celery
    #   django-celery-beat
    #   kombu
uritemplate==4.1.1
    # via drf-yasg
urllib3==2.2.3
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via
    #   blessed
    #   prompt-toolkit
wheel==0.44.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
